requirements.txt
instagrapi==2.1.2
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
pymongo==4.10.1
google-generativeai==0.8.3
openai==1.58.1
moviepy==1.0.3
